added results from google colab
Did random search in Google colab (to see whether its gpu is any good). It was trash, so bought 100 compute units and did an extensive random search in the hyperparameter space. Results are saved in the Results folder. Have to write a script to process these.
Got rid of grid search code since we will not be using it anymore.