created a script to use the mars model to get joint positions of all data
Wrote python script to iterate over each subject and each exercise and use the MARS model to predict the joint position for all the radar data. Saved results to a csv file similar to the radar data csv. The csv contains 57 columns for the 19 joint positions, and the subject and exercise label. It can now be used for feature extraction, feature selection, and ML training.